+++ app.js
@@ -1,10 +1,8 @@
-require('dotenv').config();
 const express = require('express');
 const mongoose = require('mongoose');
 const bodyParser = require('body-parser');
 const { errors } = require('celebrate');
 const cors = require('cors');
-const rateLimit = require('express-rate-limit');
 const helmet = require('helmet');
 const usersRouter = require('./routes/users');
 const moviesRouter = require('./routes/movies');
@@ -13,57 +11,28 @@ const router = require('./routes/index');
 const auth = require('./middlewares/auth');
 const { requestLogger, errorLogger } = require('./middlewares/logger');
 const errorHandler = require('./middlewares/errorHandler');
-
-const limiter = rateLimit({
-  windowMs: 15 * 60 * 1000,
-  max: 100,
-});
-
-const options = {
-  origin: [
-    'http://localhost:3001',
-  ],
-  methods: ['GET', 'HEAD', 'PUT', 'PATCH', 'POST', 'DELETE'],
-  preflightContinue: false,
-  optionsSuccessStatus: 204,
-  allowedHeaders: ['Content-Type', 'origin', 'Authorization', 'Accept'],
-  credentials: true,
-};
+const {
+  limiter, DB_NAME, DB_OPTIONS, PORT, CORS_OPTIONS,
+} = require('./utils/config');
 
 const app = express();
 
+mongoose.connect(DB_NAME, DB_OPTIONS);
+
 app.use(helmet());
-
-app.use('*', cors(options));
-
-const { PORT = 3000 } = process.env;
-const { DATA_BASE, NODE_ENV } = process.env;
+app.use(cors(CORS_OPTIONS));
 
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: true }));
-
-mongoose.connect(NODE_ENV === 'production' ? DATA_BASE : 'mongodb://localhost:27017/bitfilmsdb', {
-  useNewUrlParser: true,
-  useCreateIndex: true,
-  useFindAndModify: false,
-});
-
 app.use(requestLogger);
-
 app.use(limiter);
-
 app.use('/', router);
-
 app.use(auth);
-
 app.use('/', usersRouter);
 app.use('/', moviesRouter);
 app.use('/', errorRouter);
-
 app.use(errorLogger);
-
 app.use(errors());
-
 app.use(errorHandler);
 
 app.listen(PORT, () => {
